feat(auth): substituir react-text-mask por use-mask-input e refatorar formulários de autenticação
- Substituído o componente react-text-mask pelo use-mask-input nos formulários de login e recuperação de senha, melhorando a flexibilidade e a manutenção do código.
- Removido o componente InputPassword do diretório de nova senha e atualizado o formulário de nova senha para utilizar o novo componente de entrada de senha.
- Adicionados novos esquemas de validação para CPF/CNPJ e SMS, centralizando a lógica de validação em auth.schema.ts.
- Refatorados os formulários para utilizar o novo esquema de validação, garantindo uma experiência de usuário mais consistente e intuitiva.
- Removidos componentes e arquivos obsoletos, limpando a base de código.
