+++ apps/web/package.json
@@ -20,7 +20,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
-    "react-text-mask": "^5.5.0"
+    "use-mask-input": "^3.4.2"
   },
   "devDependencies": {
     "@coobrastur/eslint-config": "workspace:*",
@@ -31,7 +31,6 @@
     "@types/node": "^20",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
-    "@types/react-text-mask": "^5.4.14",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5",
